+++ tests/task/test_cutOutDonutsUnpairedTask.py
@@ -69,7 +69,7 @@ class TestCutOutDonutsUnpairedTask(lsst.utils.tests.TestCase):
         else:
             cls.baseRunName = "run1"
             if cls.baseRunName in collectionsList:
-                cleanUpCmd = writeCleanUpRepoCmd(cls.repoDir, cls.runName)
+                cleanUpCmd = writeCleanUpRepoCmd(cls.repoDir, cls.baseRunName)
                 runProgram(cleanUpCmd)
 
         collections = "refcats/gen2,LSSTCam/calib,LSSTCam/raw/all"
@@ -171,9 +171,14 @@ class TestCutOutDonutsUnpairedTask(lsst.utils.tests.TestCase):
         ) = self._getDataFromButler()
 
         # Test normal behavior
-        taskOut = self.task.run(
-            [copy(exposureExtra), copy(exposureIntra)],
-            [donutCatalogExtra, donutCatalogIntra],
+        taskOutExtra = self.task.run(
+            copy(exposureExtra),
+            donutCatalogExtra,
+            camera,
+        )
+        taskOutIntra = self.task.run(
+            copy(exposureIntra),
+            donutCatalogIntra,
             camera,
         )
 
@@ -185,7 +190,13 @@ class TestCutOutDonutsUnpairedTask(lsst.utils.tests.TestCase):
             "donutStamps", dataId=self.dataIdIntra, collections=[self.runName]
         )
 
+<<<<<<< HEAD
         for butlerStamp, taskStamp in zip(donutStampsExtra, taskOut.donutStamps[0]):
             self.assertMaskedImagesAlmostEqual(butlerStamp.stamp_im, taskStamp.stamp_im)  # type: ignore
         for butlerStamp, taskStamp in zip(donutStampsIntra, taskOut.donutStamps[1]):
+=======
+        for butlerStamp, taskStamp in zip(donutStampsExtra, taskOutExtra.donutStamps):
+            self.assertMaskedImagesAlmostEqual(butlerStamp.stamp_im, taskStamp.stamp_im)  # type: ignore
+        for butlerStamp, taskStamp in zip(donutStampsIntra, taskOutIntra.donutStamps):
+>>>>>>> 6962d285 (All Peter's work)
             self.assertMaskedImagesAlmostEqual(butlerStamp.stamp_im, taskStamp.stamp_im)  # type: ignore